requirements.txt
matplotlib=3.4.1
pylab=1.3.2
braces=0.2.0
djangorestframework=3.12.4
requests=2.25.1